+++ services/web/app/static/js/all.js
@@ -1,3 +1,7 @@
 $(document).ready(function(){
   $('.toast').toast('show');
+
+  $('#history').DataTable( {
+});
+
 });
