+++ services/web/app/storage/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 storage = Blueprint("storage", __name__)
 
-from .views import *
+from .routes import *
