+++ services/web/app/storage/enums.py
@@ -7,6 +7,7 @@ class FixedColdStorageType(FormEnum):
 
 
 class FixedColdStorageTemps(FormEnum):
+    L = "-150 to -86°C"
     A = "-85 to -60°C"
     B = "-59 to -35°C"
     C = "-34 to -18°C"
