+++ services/web/app/storage/routes/__init__.py
@@ -1,0 +1,6 @@
+from .misc import *
+from .site import *
+from .lts import *
+from .cryobox import *
+from .room import *
+from .shelf import *
