+++ services/web/app/storage/views/__init__.py
@@ -1,6 +1,1 @@
-from .misc import *
-from .site import *
-from .lts import *
-from .cryobox import *
-from .room import *
 from .shelf import *
