+++ services/web/app/templates/storage/shelf/view.html
@@ -5,22 +5,19 @@
           rel="stylesheet">
 {% endblock %}
 
-{% set title = "LIMBSHF-" %}
 
-{% block title %}{% endblock %}
+{% block title %}LIMBSHF-{{ shelf.id }} ({{ shelf.name }}){% endblock %}
 
 {% block body %}
     <div id="sidebar-wrapper">
         {{ storage_navbar() }}
         <div class="storage container">
 
-            <h1><span class="secondary-heading">Freezer Info Here</span>
-                <i class="fa fa-bars"></i> LIMBSHF-{{ shelf.id }}: {{ shelf.name }}
-            </h1>
+            <h1><i class="fa fa-bars"></i> LIMBSHF-{{ shelf.id }}: {{ shelf.name }}</h1>
 
 
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center">
-                <h3>Samples</h3>
+                <h2><i class="fas fa-vial"></i> Samples</h3>
                 <a href="{{ url_for('storage.assign_sample_to_shelf', shelf_id=shelf.id) }}">
                     <button class="btn btn-primary"">
                         <i class="fa fa-plus"></i>
@@ -28,38 +25,76 @@
                     </button>
                 </a>
             </div>
-            {% if samples|length > 0 %}
-                {% for sample in samples %}
-                <div class="card">
-                    <div class="card-header">
-                        {{ sample.id }}
-                    </div>
-                </div>
+
+            <table id="sampleTable" class="table table-striped table-bordered" style="width:100%">
+                <thead>
+                <tr>
+                    <th>Identifier</th>
+                    <th>Sample Type</th>
+                    <th>Sample Status</th>
+                    <th>Collection Date</th>
+                    <th>Sample Creator</th>
+                </tr>
+                </thead>
+                <tbody>
+                {% for sample_id, sample in samples.items() %}
+                    <tr>
+                        <td>
+                            <a href="{{ url_for('sample.view', sample_id=sample_id) }}">
+                                <i class="fas fa-vial"></i> LIMBSMP-{{ sample_id }}
+                            </a>
+                        </td>
+                        <td>{{ sample["sample_type"] }}</td>
+                        <td>{{ sample["sample_status"] }}</td>
+                        <td>{{ sample["creation_date"].strftime('%Y-%m-%d') }}</td>
+                        <td>
+                            {{ sample["user_information"]["email"] }}
+                        </td>
+                    </tr>
                 {% endfor %}
-            {% else %}
-                <span class="text-secondary">No Samples</span>
-            {% endif %}
+                </tbody>
+            </table>
+    
+        </div>
+    
+
 
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center">
-                <h3>Cryovial Boxes</h3>
+                <h2><i class="fa fa-cube"></i> Cryovial Boxes</h2>
                 <a href="{{ url_for('storage.add_cryobox', shelf_id=shelf.id) }}">
                     <button class="btn btn-primary"">
                         <i class="fa fa-plus"></i>
                         Add Cryovial Box
                     </button>
                 </a>
-            </div>
-            {% if cryoboxes|length > 0 %}
-                {% for box in cryoboxes %}
-                <div class="card">
-                    <div class="card-header">
-                        {{ box.serial }}
+            </div> 
+            {% if cryoboxes|length > 1 %}
+            {% for row in cryoboxes %}
+                <div class="row">
+                    {% for box in row %}
+                    <div class="col-3">
+                    <div class="card">
+                        <div class="card-header text-center">
+                            <h3><i class="fa fa-cube"></i></h3>
+                        </div>
+                        <div class="card-body">
+                            <p>Serial Number: {{ box.serial }}</p>
+                            <p>Dimensions: {{ box.num_rows }} x {{ box.num_cols }}</p>
+                        </div>
+                        <div class="card-footer text-center">
+                            <div class="btn btn-primary full-width">
+                                Go
+                            </div>
+                        </div>
                     </div>
+                    </div>
+                    {% endfor %}
                 </div>
-                {% endfor %}
-            {% else %}
-                <span class="text-secondary">No Cryovial Boxes</span>
-            {% endif %}
+            {% endfor %}
+            {%else %}
+                <span class="text-secondary">No Cryovial Boxes found</span>
+            {%endif%}
+
         </div>
     </div>
 {% endblock %}
